+++ src/onegov.form/onegov/form/__init__.py
@@ -1,11 +1,18 @@
 from onegov.form.collection import FormCollection
 from onegov.form.core import Form, with_options
-from onegov.form.models import FormDefinition, FormSubmission
+from onegov.form.models import (
+    FormDefinition,
+    FormSubmission,
+    PendingFormSubmission,
+    CompleteFormSubmission
+)
 
 __all__ = [
     'Form',
     'FormCollection',
     'FormDefinition',
     'FormSubmission',
+    'PendingFormSubmission',
+    'CompleteFormSubmission',
     'with_options'
 ]