+++ src/onegov.form/onegov/form/models.py
@@ -1,12 +1,14 @@
 from onegov.core.orm import Base
 from onegov.core.orm.mixins import TimestampMixin
-from onegov.core.orm.types import JSON
+from onegov.core.orm.types import JSON, UUID
+from onegov.form.errors import UnableToComplete
 from onegov.form.parser import parse_form
-from sqlalchemy import Column, ForeignKey, Integer, Text
+from sqlalchemy import Column, Enum, ForeignKey, Text
 from sqlalchemy.orm import (
     deferred,
     relationship,
 )
+from uuid import uuid4
 
 
 class FormDefinition(Base, TimestampMixin):
@@ -53,8 +55,8 @@ class FormSubmission(Base, TimestampMixin):
 
     __tablename__ = 'submissions'
 
-    #: internal id of the form submission
-    id = Column(Integer, primary_key=True)
+    #: id of the form submission
+    id = Column(UUID, primary_key=True, default=uuid4)
 
     #: name of the form this submission belongs to
     name = Column(Text, ForeignKey(FormDefinition.name), nullable=False)
@@ -67,8 +69,34 @@ class FormSubmission(Base, TimestampMixin):
     #: the submission data
     data = Column(JSON, nullable=False)
 
+    #: the state of the submission
+    state = Column(
+        Enum('pending', 'complete', name='submission_state'),
+        nullable=False
+    )
+
+    __mapper_args__ = {
+        "polymorphic_on": 'state'
+    }
+
     @property
     def form_class(self):
         """ Parses the form definition and returns a form class. """
 
         return parse_form(self.definition)
+
+    def complete(self):
+        """ Changes the state to 'complete', if the data is valid. """
+
+        if not self.form_class(data=self.data).validate():
+            raise UnableToComplete()
+
+        self.state = 'complete'
+
+
+class PendingFormSubmission(FormSubmission):
+    __mapper_args__ = {'polymorphic_identity': 'pending'}
+
+
+class CompleteFormSubmission(FormSubmission):
+    __mapper_args__ = {'polymorphic_identity': 'complete'}